requirements.txt
django
djangorestframework
stripe
#stripe_subscriptions